Add files via upload
In C++, a function is a block of code that performs a specific task. It is a self-contained unit of code that can be called and executed from different parts of a program. Functions in C++ provide modularity and code reusability, allowing you to break down complex tasks into smaller, manageable parts. 

ADVANTAGE-: size of executable will be reduced.
Complile time polymorphism
function overloading and operator overloading are two types of compile time polymorphism
Assign some functionality logic to an operator is called an operator overloading.
use structure to group an imaginary part together, or in a class.